+++ lib/keyStack.js
@@ -12,53 +12,54 @@
 // TODO: refactor to write more like react-side-effect and removing reliance
 //       on prop names, then setup as external lib (perferably open-source)
 
-var MAPPINGS = { // TODO: setup external json mapping file
-  27: 'esc'
+const MAPPINGS = { // TODO: setup external json mapping file
+	27: 'esc'
 }
 
-var _ancestryStack = []
+let _ancestryStack = []
 
-var keyStack = module.exports = {
+const keyStack = {
 
-  stack: function(element) {
+	stack: function(element) {
 
-    // first binding => setup
-    if (_ancestryStack.length == 0) {
-      document.addEventListener('keydown', keyStack.listener)
-    }
+		// first binding => setup
+		if (_ancestryStack.length == 0) {
+			document.addEventListener('keydown', keyStack.listener)
+		}
 
-    _ancestryStack.push(element)
-  },
+		_ancestryStack.push(element)
+	},
 
-  remove: function(element) {
+	remove: function(element) {
 
-    // remove from stack
-    var index = _ancestryStack.indexOf(this)
-    _ancestryStack.splice(index, 1)
+		// remove from stack
+		const index = _ancestryStack.indexOf(this)
+		_ancestryStack.splice(index, 1)
 
-    // no more bindings => un-setup
-    if (_ancestryStack.length == 0) {
-      document.removeEventListener('keydown', keyStack.listener)
-    }
-  },
+		// no more bindings => un-setup
+		if (_ancestryStack.length == 0) {
+			document.removeEventListener('keydown', keyStack.listener)
+		}
+	},
 
-  listener: function(event) {
+	listener: function(event) {
 
-    if (['INPUT', 'TEXTAREA'].indexOf(event.target.tagName) >= 0) {
+		if (['INPUT', 'TEXTAREA'].indexOf(event.target.tagName) >= 0) {
       return
     }
 
-    var mapping = MAPPINGS[event.keyCode]
+    const mapping = MAPPINGS[event.keyCode]
 
     for (let i = _ancestryStack.length - 1; i >= 0; --i) {
 
-      var element = _ancestryStack[i]
-      if (element.props.shortcut == mapping && element.isMounted()) { // WARNING: isMounted will be deprecated
-        element.props.action() // TODO: check if should bind original 'this'
-        return
-      }
+    	const element = _ancestryStack[i]
+    	if (element.props.shortcut == mapping && element.isMounted()) { // WARNING: isMounted will be deprecated
+    		element.props.action() // TODO: check if should bind original 'this'
+    		return
+    	}
     }
-  }
+	}
 }
 
+export default keyStack
 // WARNING: only call keyStack client-size, i.e. componentDidMount/WillUnmount
